refactor/#198: MapViewController의 마커 업데이트 로직을 MessageMarkerManager로 이동
